Added infrustructure for passing transport properties around

+++ src/gas/gas/gas_state.h
@@ -3,8 +3,6 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "Kokkos_Macros.hpp"
-
 template <typename T>
 struct GasState {
 public:
@@ -34,7 +32,7 @@ public:
 public:
     GasStates() {}
 
-    GasStates(int n) {
+    GasStates(size_t n) {
         rho_idx_ = 0;
         pressure_idx_ = 1;
         temp_idx_ = 2;
@@ -50,19 +48,21 @@ public:
           energy_idx_(3) {}
 
     KOKKOS_INLINE_FUNCTION
-    T& rho(const int cell_i) const { return data_(cell_i, rho_idx_); }
+    T& rho(const size_t cell_i) const { return data_(cell_i, rho_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    T& rho(const int cell_i) { return data_(cell_i, rho_idx_); }
+    T& rho(const size_t cell_i) { return data_(cell_i, rho_idx_); }
 
     KOKKOS_INLINE_FUNCTION
     auto rho() { return Kokkos::subview(data_, Kokkos::ALL, rho_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    T& pressure(const int cell_i) const { return data_(cell_i, pressure_idx_); }
+    T& pressure(const size_t cell_i) const {
+        return data_(cell_i, pressure_idx_);
+    }
 
     KOKKOS_INLINE_FUNCTION
-    T& pressure(const int cell_i) { return data_(cell_i, pressure_idx_); }
+    T& pressure(const size_t cell_i) { return data_(cell_i, pressure_idx_); }
 
     KOKKOS_INLINE_FUNCTION
     auto pressure() {
@@ -70,19 +70,19 @@ public:
     }
 
     KOKKOS_INLINE_FUNCTION
-    T& temp(const int cell_i) const { return data_(cell_i, temp_idx_); }
+    T& temp(const size_t cell_i) const { return data_(cell_i, temp_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    T& temp(const int cell_i) { return data_(cell_i, temp_idx_); }
+    T& temp(const size_t cell_i) { return data_(cell_i, temp_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    T& energy(const int cell_i) const { return data_(cell_i, energy_idx_); }
+    T& energy(const size_t cell_i) const { return data_(cell_i, energy_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    T& energy(const int cell_i) { return data_(cell_i, energy_idx_); }
+    T& energy(const size_t cell_i) { return data_(cell_i, energy_idx_); }
 
     KOKKOS_INLINE_FUNCTION
-    void copy_gas_state(const GasState<T>& gs, const int i) {
+    void copy_gas_state(const GasState<T>& gs, const size_t i) {
         rho(i) = gs.rho;
         pressure(i) = gs.pressure;
         temp(i) = gs.temp;
@@ -100,7 +100,7 @@ public:
     }
 
     KOKKOS_INLINE_FUNCTION
-    int size() const { return data_.extent(0); }
+    size_t size() const { return data_.extent(0); }
 
 public:
     view_type data_;
